特卖添加商品

+++ MyPlugins/Nop.Plugin.Widgets.ProductSpecialSale/Domain/SpecialSaleProduct.cs
@@ -13,8 +13,8 @@ namespace Nop.Plugin.Widgets.ProductSpecialSale.Domain
     public class SpecialSaleProduct : BaseEntity
     {
         public int ProdcutId { get; set; }
-        public double Price { get; set; }
-        public double OriginalPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal OriginalPrice { get; set; }
         public int DisplayOrder { get; set; }
         public int Quantity { get; set; }
         public SpecialSaleStage SpecialSaler { get; set; }
